xay dung chuc nang quan ly san pham (phan loai san pham)

+++ Model/SanPham.cs
@@ -11,13 +11,13 @@ namespace QuanLyCuaHangGiaDung.Model
         private string _MaSP;
         private string _TenSP;
         private string _Loai;
-        private float _GiaBan;
+        private double _GiaBan;
         private string _DVT;
 
         public string MaSP { get => _MaSP; set => _MaSP = value; }
         public string TenSP { get => _TenSP; set => _TenSP = value; }
         public string Loai { get => _Loai; set => _Loai = value; }
-        public float GiaBan { get => _GiaBan; set => _GiaBan = value; }
+        public double GiaBan { get => _GiaBan; set => _GiaBan = value; }
         public string DVT { get => _DVT; set => _DVT = value; }
 
         public SanPham() { }
